requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
pyyaml>=6.0
python-dateutil>=2.8.2
plotly>=5.15.0
altair>=5.0.0
